user system/requirements.txt
flask
flask_sqlalchemy
PyJWT
mysql-connector-python